reference/netology_pd_diplom/requirements.txt
django~=3.2.20
djangorestframework~=3.14.0
django-rest-passwordreset
psycopg2-binary
pyyaml~=6.0.1
ujson~=5.7.0
requests~=2.31.0
celery~=5.2.7